Make execute method on WASM client for compatibility with async_impl (#989)
The async_impl of reqwest has a `execute` impl on Client which is used
to execute a request and return a Result Future. When converting a crate
from async to wasm this method is missing, requiring forking and
rewriting the crate. By introducing this method less errors will be
introduced when trying to compile to wasm